封送字符串数组

+++ samples/dll/NativeLib.cpp
@@ -1,6 +1,5 @@
 #include <NativeLib.h>
 
-
 void PrintMsg(const char *msg)
 {
     printf("%s\n", msg);
@@ -173,7 +172,7 @@ void ShowNativeStructSize(size_t size)
     // 支持中文字符显示
     _wsetlocale(LC_ALL, L"chs");
 
-    wprintf(L"The size of unmanaged struct is (%d) bytes。\n", size);
+    wprintf(L"The size of unmanaged struct is (%d) bytes。\n", (int)size);
 }
 
 void __cdecl TestStructArgumentByVal(SIMPLESTRUCT simpleStruct)
@@ -287,17 +286,16 @@ void TestStructInStructByRef(PPERSON pPerson)
     const auto firstLen = strlen(pPerson->pName->first);
     const auto lastLen = strlen(pPerson->pName->last);
 
-    auto* const temp = static_cast<char*>(CoTaskMemAlloc(sizeof(char) * (firstLen + lastLen + 2)));
-	sprintf_s(temp, firstLen + lastLen + 2, "%s %s", pPerson->pName->last, pPerson->pName->first);
+    auto *const temp = static_cast<char *>(CoTaskMemAlloc(sizeof(char) * (firstLen + lastLen + 2)));
+    sprintf_s(temp, firstLen + lastLen + 2, "%s %s", pPerson->pName->last, pPerson->pName->first);
 
-	CoTaskMemFree(pPerson->pName->displayName);
-	pPerson->pName->displayName = temp;
+    CoTaskMemFree(pPerson->pName->displayName);
+    pPerson->pName->displayName = temp;
 }
 
 void TestStructInStructByVal(PPERSON2 pPerson)
 {
-	// 支持中文字符显示
-	setlocale(LC_ALL, "chs");
-	printf("姓 = %s\n名 = %s\n年龄 = %i\n\n",pPerson->name.last, pPerson->name.first, pPerson->age);
+    // 支持中文字符显示
+    setlocale(LC_ALL, "chs");
+    printf("姓 = %s\n名 = %s\n年龄 = %i\n\n", pPerson->name.last, pPerson->name.first, pPerson->age);
 }
-
